tweaks: add toast to reset password and change devTool on staging
