+++ src/App.jsx
@@ -2,32 +2,32 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { Provider } from "react-redux";
 import store from "./redux/store";
 import Home from "./pages/Home";
-import Login from "./pages/Login";
-import Register from "./pages/Register";
+import Login from "./pages/Auth/Login";
+import Register from "./pages/Auth/Register";
 import NotFound from "./pages/NotFound";
-import ResetPassword from "./pages/ResetPassword";
+import ResetPassword from "./pages/Auth/ResetPassword";
 import SendReset from "./components/Auth/SendReset";
 function App() {
-  return (
-    <>
-      <Provider store={store}>
-        <BrowserRouter>
-          <Routes>
-            <Route path="/" element={<Home />} />
+    return (
+        <>
+            <Provider store={store}>
+                <BrowserRouter>
+                    <Routes>
+                        <Route path="/" element={<Home />} />
 
-            {/* Authorization */}
-            <Route path="/login" element={<Login />} />
-            <Route path="/register" element={<Register />} />
-            <Route path="/reset" element={<SendReset />} />
-            <Route path="/forgot/:id" element={<ResetPassword />} />
+                        {/* Authorization */}
+                        <Route path="/login" element={<Login />} />
+                        <Route path="/register" element={<Register />} />
+                        <Route path="/reset" element={<SendReset />} />
+                        <Route path="/forgot/:id" element={<ResetPassword />} />
 
-            {/* Handle Pages Not Found */}
-            <Route path="*" element={<NotFound />} />
-          </Routes>
-        </BrowserRouter>
-      </Provider>
-    </>
-  );
+                        {/* Handle Pages Not Found */}
+                        <Route path="*" element={<NotFound />} />
+                    </Routes>
+                </BrowserRouter>
+            </Provider>
+        </>
+    );
 }
 
 export default App;
