+++ src/redux/store.js
@@ -2,6 +2,6 @@ import { configureStore } from "@reduxjs/toolkit";
 import rootReducers from "./reducers";
 
 export default configureStore({
-  reducer: rootReducers,
-  devTools: import.meta.env.MODE === "development",
+    reducer: rootReducers,
+    devTools: import.meta.env.VITE_MODE !== "production",
 });
